# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Plutonium SMP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /store/items:
    get:
      operationId: getStoreItems
      tags: [store]
      summary: Get all store items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Store items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreItem"
  /store/items/{id}:
    get:
      operationId: getStoreItem
      tags: [store]
      summary: Get store item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
  /store/purchase:
    post:
      operationId: purchaseItem
      tags: [purchases]
      summary: Purchase a store item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequest"
      responses:
        "200":
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: Get user's tickets
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "200":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketWithMessages"
  /tickets/{id}/messages:
    post:
      operationId: sendTicketMessage
      tags: [tickets]
      summary: Send a message in a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
  /tickets/{id}/close:
    post:
      operationId: closeTicket
      tags: [tickets]
      summary: Close a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/store/items:
    post:
      operationId: adminCreateStoreItem
      tags: [admin]
      summary: Create a store item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreItemRequest"
      responses:
        "200":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
  /admin/store/items/{id}:
    put:
      operationId: adminUpdateStoreItem
      tags: [admin]
      summary: Update a store item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
    delete:
      operationId: adminDeleteStoreItem
      tags: [admin]
      summary: Delete a store item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    StoreItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [ranks, crate_keys, coins, cosmetics]
        price:
          type: integer
        currency:
          type: string
          enum: [owo, usd]
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        badge:
          type: string
        badgeColor:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - category
        - price
        - currency
        - isActive
        - isFeatured
        - createdAt
    CreateStoreItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: integer
        currency:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        badge:
          type: string
        badgeColor:
          type: string
      required:
        - name
        - description
        - category
        - price
        - currency
    PurchaseRequest:
      type: object
      properties:
        itemId:
          type: string
        couponCode:
          type: string
      required:
        - itemId
    Purchase:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        itemId:
          type: string
        itemName:
          type: string
        itemCategory:
          type: string
        pricePaid:
          type: integer
        currency:
          type: string
        couponUsed:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        createdAt:
          type: string
      required:
        - id
        - userId
        - itemId
        - itemName
        - itemCategory
        - pricePaid
        - currency
        - status
        - createdAt
    Ticket:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        subject:
          type: string
        category:
          type: string
          enum: [support, purchase_issues, report_player, general]
        status:
          type: string
          enum: [open, pending, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - subject
        - category
        - status
        - priority
        - messageCount
        - createdAt
        - updatedAt
    TicketMessage:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        userId:
          type: string
        username:
          type: string
        role:
          type: string
        content:
          type: string
        isStaff:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - userId
        - username
        - content
        - isStaff
        - createdAt
    TicketWithMessages:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/Ticket"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
      required:
        - ticket
        - messages
    CreateTicketRequest:
      type: object
      properties:
        subject:
          type: string
        category:
          type: string
        priority:
          type: string
        message:
          type: string
      required:
        - subject
        - category
        - message
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...